+++ cmd/get/nodepools/printer.go
@@ -32,9 +32,18 @@ func (r *runner) printOutput(npResource nodepool.Resource) error {
 		case key.ProviderCAPA:
 			capabilities := feature.New(feature.ProviderCAPA)
 			resource = provider.GetCAPATable(npResource, capabilities)
+		case key.ProviderCAPZ:
+			capabilities := feature.New(feature.ProviderCAPZ)
+			resource = provider.GetCAPITable(npResource, capabilities)
 		case key.ProviderAzure:
 			capabilities := feature.New(feature.ProviderAzure)
 			resource = provider.GetAzureTable(npResource, capabilities)
+		case key.ProviderCloudDirector:
+			capabilities := feature.New(feature.ProviderCloudDirector)
+			resource = provider.GetCAPITable(npResource, capabilities)
+		case key.ProviderVSphere:
+			capabilities := feature.New(feature.ProviderVSphere)
+			resource = provider.GetCAPITable(npResource, capabilities)
 		}
 
 		printOptions := printers.PrintOptions{