+++ pkg/data/domain/nodepool/common.go
@@ -48,6 +48,21 @@ func (s *Service) getByName(ctx context.Context, provider, name, namespace, clus
 			if err != nil {
 				return nil, microerror.Mask(err)
 			}
+		case key.ProviderCAPZ:
+			np, err = s.getByIdCAPI(ctx, name, namespace, clusterName)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
+		case key.ProviderCloudDirector:
+			np, err = s.getByIdCAPI(ctx, name, namespace, clusterName)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
+		case key.ProviderVSphere:
+			np, err = s.getByIdCAPI(ctx, name, namespace, clusterName)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
 		default:
 			return nil, microerror.Mask(invalidProviderError)
 		}
@@ -77,6 +92,24 @@ func (s *Service) getAll(ctx context.Context, provider, namespace, clusterID str
 
 		case key.ProviderCAPA:
 			npCollection, err = s.getAllCAPA(ctx, namespace, clusterID)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
+
+		case key.ProviderCAPZ:
+			npCollection, err = s.getAllCAPI(ctx, namespace, clusterID)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
+
+		case key.ProviderCloudDirector:
+			npCollection, err = s.getAllCAPI(ctx, namespace, clusterID)
+			if err != nil {
+				return nil, microerror.Mask(err)
+			}
+
+		case key.ProviderVSphere:
+			npCollection, err = s.getAllCAPI(ctx, namespace, clusterID)
 			if err != nil {
 				return nil, microerror.Mask(err)
 			}